+++ app/src/main/java/first/android/cis/domain/models/news/NewsList.kt
@@ -1,3 +1,3 @@
-package first.android.cis.models
+package first.android.cis.domain.models
 
 class NewsList : ArrayList<NewsListItem>()
