+++ app/src/main/java/first/android/cis/domain/models/user/UserInfoList.kt
@@ -1,0 +1,3 @@
+package first.android.cis.domain.models.users
+
+class UserInfoList : ArrayList<UserInfo>()
