--- app/src/main/java/first/android/cis/models/users/UserInfoList.kt
@@ -1,3 +1,0 @@
-package first.android.cis.models.users
-
-class UserInfoList : ArrayList<UserInfo>()
