+++ app/src/main/java/first/android/cis/network/services/SignInService.kt
@@ -1,4 +1,4 @@
-package first.android.cis.ui.signUpIn
+package first.android.cis.presentation.signUpIn.signIn
 
 import android.content.Context
 import android.widget.Button
@@ -6,7 +6,6 @@ import android.widget.Toast
 import androidx.fragment.app.FragmentActivity
 import androidx.navigation.NavDirections
 import androidx.navigation.findNavController
-import first.android.cis.MainActivity
 import first.android.cis.models.users.AuthData
 import first.android.cis.models.users.UserToken
 import first.android.cis.network.Retrofit
