+++ app/src/main/java/first/android/cis/presentation/news/openedNews/OpenedNewsViewModel.kt
